aider: Added play and select buttons to customOverlay in createGalleryItems function.
# Aider chat conversation:

USER: hnadleVideoSelect doesnt seem to be working after that change, i want 2 custom bottons on them , one to play and one to select

ASSISTANT: To add two custom buttons, one for playing the video and another for selecting it, you can modify the `customOverlay` property in the `createGalleryItems` function. Here's the *SEARCH/REPLACE block* to make the changes:

nodejs/pages/videos.js
```javascript
<<<<<<< SEARCH
      customOverlay: (
        <div style={{ position: 'absolute', bottom: 0, right: 0 }}>
          <button onClick={() => onSelect(videoId)}>Select</button>
        </div>
      ),
=======
      customOverlay: (
        <div style={{ position: 'absolute', bottom: 0, right: 0 }}>
          <button onClick={() => handleVideoSelect(videoId)}>Select</button>
          <button onClick={() => setSelectedVideo(videoId)}>Play</button>
        </div>
      ),
>>>>>>> REPLACE
```

This change will add a "Play" button next to the "Select" button. When the "Play" button is clicked, the video will be played below the gallery.