aider: Added "Update XML" button with onClick event above Monaco editor.
# Aider chat conversation:

USER: add a button above the monaco editor, should display text "Update XML" and should have a onclick.

ASSISTANT: To add a button above the Monaco editor with the text "Update XML" and an onClick event, we need to modify `nodejs/components/AssetsAggregation.js`:

1. Add the button element above the Monaco editor.
2. Add an onClick event handler to the button.

Here is the *SEARCH/REPLACE* block:

nodejs/components/AssetsAggregation.js
```javascript
<<<<<<< SEARCH
      {xmlCode && (
=======
      <button
        className="bg-blue-500 text-white px-4 py-2 rounded-lg mb-4"
        onClick={() => {
          console.log("Update XML button clicked");
        }}
      >
        Update XML
      </button>
      {xmlCode && (
>>>>>>> REPLACE
```

This change will add the "Update XML" button above the Monaco editor and log a message to the console when the button is clicked. You can replace the `console.log` statement with your desired functionality.